src/valuation/requirements.txt
fastapi
uvicorn[standard]
gspread
google-auth
google-auth-oauthlib
google-api-python-client
requests
pandas
prophet 
numpy
scipy
matplotlib
scikit-learn

polygon-api-client==1.14.4
pydantic==2.10.6
python-dotenv
uv==0.6.16 